Further destructuring of components and styling.
