+++ src/App.js
@@ -55,6 +55,7 @@ function App() {
             <Route exact path="/resources" component={Resources} />
             <Route exact path="/report" component={Report} />
           </Switch>
+          <nav></nav>{/* Navigation Placeholder */}
         </div>
       </Router>
     </Provider>
