+++ src/App.scss
@@ -8,6 +8,9 @@
 @media (max-width: 505px) {
   .App { 
     width: 100%;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
    }
   .Wipe { display: none; }
 }
@@ -29,6 +32,5 @@ button:focus {
   outline: none;
   border: none;
 }
-
