+++ src/components/activity/Activity.js
@@ -1,18 +1,22 @@
 import React from "react";
 import Filters from "./Filters";
-import List from "./results/List";
-import Sort from "./results/Sort";
-// import Map from "./map/Map";
+// import List from "./results/List";
+
+import SearchBar from "./search/SearchBar";
+import ResultsContainer from "./results/ResultsContainer";
+// import Sort from "./results/Sort";
+import MapContainer from "./map/MapContainer";
 
 const Activity = () => {
   return (
-    <>
-      {/* Search Bar */}
+    <div className="main-container no-scroll">
+      <SearchBar />
       <Filters /> {/* As a popup menu that overlays the list when open but not the search bar */}
-      <Sort />{/* Place this inside of the list at the very top */}
-      <List />
-      {/* Map */}
-    </>
+      {/* <Sort />Place this inside of the list at the very top */}
+      {/* <List /> */}
+      <ResultsContainer />
+      <MapContainer /> 
+    </div>
   );
 };
