+++ src/components/activity/map/MapContainer.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const MapContainer = () => {
+    return (
+        <div className="map-container">
+            
+        </div>
+    )
+}
+
+export default MapContainer;
