+++ src/styling/activity/card.scss
@@ -8,9 +8,9 @@
   justify-content: flex-start;
 
   width: 100%;
-  min-height: calc(100vh - 78px);
+  height: calc(100% - 30px);
 
-  margin: 78px 0 0 0;
+  margin: 0;
   position: relative;
 }
