+++ src/styling/activity/map.scss
@@ -1,0 +1,12 @@
+@import "../variables.scss";
+
+.map-container {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    top: 0;
+    z-index: 998;
+    background: rgba(51, 51, 51, 0.425);
+    border: 5px solid $font-color;
+    display: block;
+}
