+++ src/styling/header.scss
@@ -18,7 +18,7 @@
 }
 
 .header {
-  position: fixed;
+  // position: fixed;
   display: flex;
   justify-content: space-between;
   align-items: center;
